[front] - refactor: rename utils.ts to types.ts across components
 - Changed the import paths from @app/components/providers/utils to @app/components/providers/types for clarity.
 - Renamed the file `utils.ts` to `types.ts` to better represent its content.
 - Updated references in various components to align with the renamed file.
