+++ front/pages/api/poke/templates/[tId].ts
@@ -7,7 +7,7 @@ import { isLeft } from "fp-ts/lib/Either";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { USED_MODEL_CONFIGS } from "@app/components/providers/utils";
+import { USED_MODEL_CONFIGS } from "@app/components/providers/types";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { TemplateResource } from "@app/lib/resources/template_resource";
 import { apiError, withLogging } from "@app/logger/withlogging";
